requirements.txt
# Pointing Gesture Analysis Pipeline - Requirements

# ⚠️ IMPORTANT: Python Version Compatibility
# MediaPipe requires Python 3.8 - 3.12 (NOT 3.13+)
# If you have Python 3.13, create a virtual environment with Python 3.11 or 3.12:
#   python3.12 -m venv venv
#   source venv/bin/activate
#   pip install -r requirements.txt

# Core dependencies
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Step 0: Data Loading & Target Detection
pyrealsense2>=2.54.0  # Optional: For live RealSense camera capture
ultralytics>=8.0.0     # YOLOv8 for cup/target detection
torch>=2.0.0           # Required by YOLO
torchvision>=0.15.0    # Required by YOLO

# Step 2: Skeleton Extraction
mediapipe>=0.10.0  # Human pose detection
scipy>=1.10.0      # For plane fitting and spatial calculations

# Step 3: Subject Detection
deeplabcut>=3.0.0  # Dog skeleton detection using SuperAnimal Quadruped model

# Visualization
matplotlib>=3.7.0  # For 2D/3D plotting
open3d>=0.17.0     # Optional: For point cloud visualization

# Data handling
pandas>=2.0.0      # For CSV export and data processing

# GUI - tkinter is included with most Python installations on macOS
# If tkinter is not available, install Python from python.org

# Development (Optional)
pytest>=7.0.0  # For testing
